Add Fcf.Class.Functor
